Modularize headline transform code
